# src/routes/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "sobermart",
      "contact": {}
    },
    "host": "kuro.asrofur.me",
    "basePath": "/sober/api",
    "securityDefinitions": {},
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/{main}{user}/admin/add-admin": {
        "post": {
          "summary": "AddAdmin",
          "tags": [
            "Admin"
          ],
          "operationId": "AddAdmin",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "first_name",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "last_name",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "username",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "email",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "password",
              "in": "formData",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            },
            {
              "name": "repassword",
              "in": "formData",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            },
            {
              "name": "main",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "user",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": []
        }
      },
...
      "/{main}{user}/admin/update": {
        "post": {
          "summary": "updateProfile",
          "tags": [
            "Admin"
          ],
          "operationId": "updateProfile",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            },
            {
              "name": "main",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "user",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
...
      "/": {
        "get": {
          "summary": "GetAllReview",
          "tags": [
            "Admin"
          ],
          "operationId": "GetAllReview",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "delete": {
          "summary": "DeleteReview",
          "tags": [
            "Admin"
          ],
          "operationId": "DeleteReview",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "put": {
          "summary": "UpdateShipment",
          "tags": [
            "Admin"
          ],
          "operationId": "UpdateShipment",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "post": {
          "summary": "UpdateProfile",
          "tags": [
            "User"
          ],
          "operationId": "UpdateProfile",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
...
      "/{main}{user}/change-pw": {
        "post": {
          "summary": "ForgotPassword",
          "tags": [
            "User"
          ],
          "operationId": "ForgotPassword",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "repassword",
              "in": "formData",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            },
            {
              "name": "password",
              "in": "formData",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            },
            {
              "name": "oldpassword",
              "in": "formData",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            },
            {
              "name": "main",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "user",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": []
        }
      },
...
      "/{main}/transaction/vendor/returns": {
        "get": {
          "summary": "OrderReturn",
          "tags": [
            "Vendor"
          ],
          "operationId": "OrderReturn",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "limit",
              "in": "query",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "main",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": []
        }
      },
...
    },
    "security": [],
...
  }